pantryhub-backend/requirements.txt
alembic==1.16.1
blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
flask-cors
python-dotenv
psycopg2-binary
firebase-admin==6.5.0
pyjwt
gunicorn
flask_apscheduler
supabase
Flask-SocketIO==5.3.6
gevent>=1.4.0
